refactor: ExitException.class 패키지 변경
