+++ src/main/java/chess/exception/ExitException.java
@@ -1,4 +1,4 @@
-package chess.domain;
+package chess.exception;
 
 public class ExitException extends RuntimeException {
     public ExitException(final String message) {
