+++ src/test/java/chess/domain/SquareTest.java
@@ -3,6 +3,7 @@ package chess.domain;
 import chess.domain.RuleImpl.King;
 import chess.domain.RuleImpl.Rook;
 import chess.domain.RuleImpl.Rule;
+import chess.exception.ExitException;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
